Update ENS.sol
The contract has an owner. And only this owner can withdraw money from the contract.

A variable with the price of one year of registration has been added to the contract and a setter of this price has been made, which can only be called by the owner.

The function has been improved, which records the structure in relation to the string.

The function now takes as an argument for how many years the user registers the domain. The minimum value is 1, the maximum is 10.
And other new additions